kvm: redefine vcpu_printf() to release vmcs
currently it is illegal to call vcpu_printf() and hvm_printf() inside a
vcpu_load()/vcpu_put() pair, since the printf functions may sleep and
vcpu_load() relies on the task not to be scheduled until vcpu_put().  so
define vcpu_printf() to release the vmcs while writing, and then reload it.

the new rules are:

  hvm_printf(): must be called outside vcpu_load()/vcpu_put()
  vcpu_printf(): must be called inside vcpu_load()/vcpu_put()