kvm single-qemu-binary: replaced all env->use_kvm with kvm_allowed
There is no need for per-cpu instance of use_kvm. kvm is either enabled or not.
We use one global variable, kvm_allowed, to represent it.

By default if USE_KVM is defined kvm is enabled (kvm_allowed==1)
One can disable kvm by adding -no-kvm to the command-line arguments.
Naturally if USE_KVM is not defined kvm is disabled.