kvm: fix bogus pagefault on writable pages
if a page is marked as dirty in the guest pte, set_pte_common() can set the
writable bit on newly-instantiated shadow pte.  this optimization avoids
a write fault after the initial read fault.

however, if a write fault instantiates the pte, fix_write_pf() incorrectly
reports the fault as a guest page fault, and the guest oopses on what appears
to be a correctly-mapped page.

fix is to detect the condition and only report a guest page fault on a user
access to a kernel page.

with the fix, a kvm guest can survive a whole night of running the kernel
hacker's screensaver (make -j9 in a loop).