kvm: qemu: synchronous signal handling
in order to ensure that signals are handled with the global lock held, we
block them, and dequeue them with sigtimedwait().  we can then call the
handlers directly.

this ensures that the device model and internal qemu data structures are not
corrupted by cpu 0 thread handling a signal while cpu 1 is servicing a pio
or mmio request.