qemu: reset more state after CR/LF
After receiving CR or LF on the monitor, reset term_last_cmd_* as
well.  This matches the setup that is done by term_show_prompt(), and is
for the case when readline_start() is not called before the next
readline_handle_byte().  Without this, unnecessary terminal escape
sequences will be printed by term_update() before the next prompt.

Signed-off-by: Jim Paris <jim@jtan.com>
Signed-off-by: Avi Kivity <avi@qumranet.com>