kvm: qemu: make migration handle errors a little better
If the remote side wasn't set up properly (i.e. the memory size didn't match),
the migration would fail, but the source host didn't actually react properly,
so the migration would still be "in-progress".  This patch fixes up the error
checking to properly quit when the remote side isn't prepared for whatever
reason.

Signed-off-by: Chris Lalancette <clalance@redhat.com>
Signed-off-by: Avi Kivity <avi@qumranet.com>