kvm: external module: preempt emulation miscompile fix
I tracked down the reason of my kernel crashing builds with the latest
kvm-userland package. It has taken a bit of time because with certain
workloads this wouldn't crash. I've been testing stuff with preempt
emulation enabled most of the time to be sure it worked (and to fix
the host breakpoint too). This will fix it by importing the task
struct before we define CONFIG_PREEMPT_NOTIFIERS. preempt.h is most
certainly unnecessary but it's a good rule of thumb to include
anything that uses CONFIG_PREEMPT_NOTIFIERS before defining it.

Signed-off-by: Andrea Arcangeli <andrea@qumranet.com>
Signed-off-by: Avi Kivity <avi@qumranet.com>