kvm: qemu: Allow kvm to use more than 4 VCPUs
This patch actually allows KVM to be used with more than 4 VCPUs.  The change
in qemu-kvm.c was pretty difficult to find because it was using an open coded
array size of 4.  I changed that array to be 256 since that's the real maximum
on x86 and the additional storage space is a pretty trivial cost.

Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>
Signed-off-by: Avi Kivity <avi@qumranet.com>