kvm: libkvm: Move kvm_get_pit() to libkvm.c common code
This fixes compilation issues for PowerPC and other non x86 archs that do not
have in kernel pit. The pit code is added into the kvm_context in kvm-common.h
and the error causing the issue is coming from a definition in qemu. This seems
to be the proper fix as there is also a common function:

  kvm_irqchip_in_kernel

for in kernel irq that handles this the same way.

Signed-off-by: Jerone Young <jyoung5@us.ibm.com>
Signed-off-by: Avi Kivity <avi@qumranet.com>