kvm: qemu: fix vmmouse with -smp
The vmport code is very broken for SMP guests.  It uses a global CPUState
that's initialized multiple times?  At any rate, since it needs to know CPU
registers for the current CPU in a PIO handler, it needs to use cpu_single_env.

This patch makes vmmouse when using -smp > 1

Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>
Signed-off-by: Avi Kivity <avi@qumranet.com>