kvm: external module: fix build on 2.6.26-rc
after updating kvm-userland.git, kvm.git and linux-2.6-hg, and after
make distclean and rebuild with slightly reduced .config, I can't
compile the external module anymore. Looking into it with V=1, $(src)
defines to "" and including /external-module-compat.h clearly fails. I
fixed it like below, because it seems more consistent to enforce the
ordering of the "special" includes in the same place, notably
$(src)/include is already included as $LINUX at point 1 of the
comment, so this looks a cleanup of superflous line in Kconfig besides
fixing my compile by moving the external-module-compat in the same
place with the other includes where `pwd` works instead of $(src) that
doesn't work anymore for whatever reason.

Signed-off-by: Andrea Arcangeli <andrea@qumranet.com>
Signed-off-by: Avi Kivity <avi@qumranet.com>