kvm: qemu: Add support for signalfd()
This patch adds compatibility code so that we can use signalfd() within QEMU.
signalfd() provides a mechanism to receive signal notification through a
file descriptor.  This is very useful in eliminating the signal/select race
condition.

If signalfd() isn't available, we spawn a thread that uses sigwaitinfo() to
emulate it.

Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>
Signed-off-by: Avi Kivity <avi@qumranet.com>