kvm: user: ppc: remove build flags
Remove -m32 build flag. Although necessary when building for a 32-bit target
with a biarch toolchain that defaults to 64-bit output, it naturally breaks
64-bit builds.

Also remove -D__powerpc__. I'm not sure why I added that in the first place;
that is automatically defined by GCC (for both 32- and 64-bit targets).

Signed-off-by: Hollis Blanchard <hollisb@us.ibm.com>
Signed-off-by: Avi Kivity <avi@redhat.com>