Make KVM slot management more robust
KVM keeps track of physical memory based on slots in the kernel.  The current
code that translates QEMU memory mappings to slots work but is not robust
in the fact of reregistering partial regions of memory.

This patch does the right thing for reregistering partial regions of memory.  It
also prevents QEMU from using KVM private slots.

Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>