Make KVMSlot a real structure
struct kvm_userspace_memory_region does not use QEMU friendly types to 
define memory slots.  This results in lots of ugly casting with warnings 
on 32-bit platforms.

This patch introduces a proper KVMSlot structure that uses QEMU types to
describe memory slots.  This eliminates many of the casts and isolates 
the type conversions to one spot.

Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>