kvm: qemu: Use guards in virtio-net for easier upstream merging (v2)
This gets virtio-net into an upstream acceptable state.  This includes
introducing USE_KVM guards for IO thread notification (where did
qemu_service_io() go?).  It also includes introducing TAP_VNET_HDR which is for
code that relies on the tap vnet support that is not currently in upstream QEMU.

Finally, it drops packets if not ready to receive.  This is unnecessary in
kvm-userspace but necessary in QEMU.

Since v1, I enable tap vnet by default, and pass the buffer size to
virtio_net_receive() as suggested by Mark.

Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>
Signed-off-by: Avi Kivity <avi@redhat.com>