kvm: testsuite: Add VMMCALL hypercall test
By default, hypercalls issue the VMCALL instruction, which is only supported
on Intel. On AMD, we detect this and patch the call site with VMMCALL, which
is only supported on AMD. When the code issues VMMCALL again, the hypercall
is called.

This works fine on both Intel as well as AMD.

When live-migrating guests from AMD machines to Intel machines, the guest
issues a VMMCALL, which should then be patched to VMCALL. Add a test for
this behaviour.

Signed-off-by: Amit Shah <amit.shah@redhat.com>
Signed-off-by: Avi Kivity <avi@redhat.com>