Make sure ACPI structures are byte-aligned
Apparently, guests are very tolerant of corrupt ACPI tables because our 
tables have been badly corrupted for some time now.  A version of 
Knoppix using a 2.6.11 kernel refused to boot and it turned out it was 
due to the interrupt override table introduced by the recent HPET 
commit.

This patch updates the BIOS and introduces a patch to pack the ACPI 
tables.  If you have a guest that used to work and is broken by the this 
commit, let me know.  We have some weird hacks in the tables that I 
suspect are work arounds for this bug.

Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>