kvm-userspace: set pci mem to start at 0xc100000 and vesa to 0xc000000
This patch make the pci mem region larger (1 giga now).
this is needed for pci devices that require large amount of memory
such as video cards.

for pea guests this patch is not an issue beacuse the guest OS will map
the rest of the ram after 0x100000000...,
for 32bits that arent pea, it mean the maximum memory that would be
avaible now is 3giga.

Signed-off-by: Izik Eidus <ieidus@redhat.com>
RH-Upstream-status: submitted(qemu-devel), waiting_review
RH-Pointer: http://article.gmane.org/gmane.comp.emulators.kvm.devel/27258