qemu: notify (async) management when migration finished at destination
SolidICE wants to know when a VM has migrated to another node.
After that event happens, it takes down the source VM and informs the
administrator of a successful migration.

The source can ask for "info migration".
The destination waits for an asynchronous notification.

This patch provides that notification by emitting
a message to qemu monitor, telling vdsm of the
successful completion of the migration process.

Remember to "notify migration on" on destination before
migration starts.

This is a modification of the patch Dan Kenigsberg sent to the list,
making use of Amit's asynchronous notifications.

Signed-off-by: Uri Lublin <uril@redhat.com>
Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>
RH-Type: improvement
RH-Upstream-status: pending