qemu: drive_init: Don't try to read passwords before monitor setup
drive_init() calls qemu_key_check(), but it doesn't make sense
to call it before the monitor was set up. This makes
qemu hang forever waiting for a password on a monitor
that doesn't exist.

Removing that call will make the password to be read at
a proper time, at read_passwords().

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>
RH-Upstream-status: submitted(qemu-devel)