host_device_remove: remove incorrect check for device name
There is no need to check for valid prefixes on the the device name
when removing them. If they are found on the client list of a vlan,
it is a valid name, regardless of the prefix used on its name.

To reproduce the bug, just run this on the monitor:

 (qemu) host_net_add user name=foobar
 (qemu) host_net_remove 0 foobar
 invalid host network device foobar
 (qemu)

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>
Bugzilla: 491156
RH-Upstream-status: submitted(qemu-devel)
Acked-by: Marcelo Tosatti <mtosatti@redhat.com>
Acked-by: Juan Quintela <quintela@redhat.com>