Fix segfault after tap host_net_remove
https://bugzilla.redhat.com/491150

If you remove a tap host device, then the associated VLANClientState
is freed, but we never run the down script, close the file descriptor
or free the TAPState memory.

The segfault here:

  772         for (vc = s->vc->vlan->first_client; vc; vc = vc->next) {

is due to that we haven't removed the tap_can_send() fd handler and
the VLANClientState has already freed.

Another obvious side effect of this bug is that the tap device still
exists (e.g. look at ifconfig -a) after host_net_remove.

This has been fixed upstream by:

  http://git.kernel.org/?p=virt/kvm/qemu-kvm.git;a=commitdiff;h=a34b6eb77

I've extracted the parts needed to cleanup the tap device and sort out
the net_cleanup() function. I've also not added cleanup as an argument
to qemu_new_vlan_client() so as to avoid changing each of the twenty
callers. The intention is to make the patch easier to review, while
statying as close as possible to upstream.

Signed-off-by: Mark McLoughlin <markmc@redhat.com>
Message-Id: <1242387172.25932.5.camel@blaa>
Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>
Bugzilla: 501729
RH-Upstream-status: not-applicable
Acked-by: Glauber Costa <glommer@redhat.com>
Acked-by: Eduardo Habkost <ehabkost@redhat.com>
Acked-by: Juan Quintela <quintela@redhat.com>