ksm ?: remove full swapping support for kvm mmu pages under ovirt (rhel 5.3)
There is problem with the anon_vma lock, as some times the kprobe run whle this
lock is token (swap path) and sometimes it run while it isnt token (page fault path)

There are ways to fix this issue, but they tend to be very complicated / risky,
condiering the fact that in rhel 5.4 we can swap using mmu notifiers, it better
to kick this code as it will reuqire alot of thinking if the fix for it is safe.

note: this doesnt effect ksm, ksm pages when they are break are still swappable
this code even tougth it was inside wp_notifier.c (belong to ksm) have nothing to
do with ksm.

Signed-off-by: Izik Eidus <ieidus@redhat.com>
Message-ID: <4A11BD10.3060905@redhat.com>
Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>
RH-Upstream-status: not-applicable
Acked-by: Andrea Arcangeli <aarcange@redhat.com>
Acked-by: john cooper <john.cooper@redhat.com>
Acked-by: Juan Quintela <quintela@redhat.com>
Bugzilla: 503793