ksm: remove kprobes usage for rhel 5.4 (v3)
Considering the fact that rhel 5.4 have mmu notifiers support, we dont need
the kprobes to handle the notifications to kvm.

Leaving the current code as it is - is not a bug, it is just matther of
unneeded code that hurt the performence for rhel 5.4 when ksm is loaded
(the copy on write performence)

(Btw, i have faked here the signeture of Chris, hope he will forgive me :))

Thanks.

Message-ID: <20090602002240.GL1494@x200.localdomain>
Obsoletes: <1243894307-17932-2-git-send-email-ieidus@redhat.com>
Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>
RH-Upstream-status: not-applicable
Acked-by: Izik Eidus <ieidus@redhat.com>
Bugzilla: 504237
Acked-by: Marcelo Tosatti <mtosatti@redhat.com>
Signed-off-by: Izik Eidus <ieidus@redhat.com>
Message-Id: <1244142579-3405-2-git-send-email-ieidus@redhat.com>
Acked-by: Andrea Arcangeli <aarcange@redhat.com>
Acked-by: john cooper <john.cooper@redhat.com>