KSM: avoid losing (but no mem leak) rmap_items inside ksm data strctures
The implication of this should be minor (worst case is to lose
visibility on some pages in stable tree, no crash, no memleak, just no
merging of pages).

The tree_item->rmap_item is the head of the list and as such it must
not be overwritten except in the case that the element we removed
(rmap_item) was the previous head of the list, in which case it would
also have rmap_item->prev set to null. This was only found by reading
source while fixing the crash with slab debug enabled. For rhev/master
and rhel5/master.

Signed-off-by: Andrea Arcangeli <aarcange@redhat.com>
Signed-off-by: Izik Eidus <ieidus@redhat.com>
Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>
RH-Upstream-status: not-applicable
Message-Id: <1243894307-17932-5-git-send-email-ieidus@redhat.com>
Message-Id: <1244142579-3405-4-git-send-email-ieidus@redhat.com>
Acked-by: Andrea Arcangeli <aarcange@redhat.com>
Acked-by: Marcelo Tosatti <mtosatti@redhat.com>
Acked-by: john cooper <john.cooper@redhat.com>
Bugzilla: 504237