Correctly assign apic ids for hotplugged cpus
BIOS assumes that apic ids are continuous from 0 to max cpus when it
builds its acpi/mptable tables, but QEMU allocates one id for ioapic,
so hot plugged cpus are created with apic id skewed by one. OS cannot
communicate with newly created cpu since the apic id of new cpu is not
the same as what ACPI tells to an OS. Do not allocate apic id to ioapic.
This shouldn't be done by QEMU anyway. This is BIOS or OS job.

Upstream status: not needed. The code was rearranged and this line is not
                 exist any more.
BZ: 504581 (part of it. Should fix f10. rhel5.3 does not support cpu hot-plug)

Signed-off-by: Gleb Natapov <gleb@redhat.com>
Message-ID: <20090616084444.GQ19508@redhat.com>
Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>
RH-Upstream-status: not-needed
Acked-by: Glauber Costa <glommer@redhat.com>
Acked-by: Avi Kivity <avi@redhat.com>
Acked-by: Don Dutile <ddutile@redhat.com>
Bugzilla: 504237