net: add '-net tap,sndbuf=' option with a sensible default
https://bugzilla.redhat.com/508861

In a bridged configuration, we should avoid overflowing the physical
NIC's TX queue and dropping packets.

The 5.4 kernel adds a new TUNSETSNDBUF ioctl() which allows a send
buffer limit for the tap device to be specified. When this limit is
reached, a tap write() will return EAGAIN and poll() will indicate the fd
isn't writable. See:

  https://bugzilla.redhat.com/495863

This allows people to tune their setups so as to avoid e.g. UDP packet
loss when the sending application in the guest out-runs the NIC in the
host.

For best effect, sndbuf= should be set to just below the capacity of
the physical NIC.

Setting it higher will cause packets to be dropped before the limit
is hit. Setting it much lower will not cause any problems unless
you set it low enough such that the guest cannot queue up new packets
before the NIC has emptied its queue.

In Linux, txqueuelen=1000 by default for ethernet NICs. Given a 1500
byte MTU, 1Mb is a good choice for sndbuf.

If it turns out that txqueuelen is actually much lower than this, then
sndbuf is essentially disabled. In the event that txqueuelen is much
higher, it's unlikely that the NIC will be able to empty a 1Mb queue.

Also, note that when using a bridge with netfilter enabled, we currently
never get EAGAIN because netfilter causes the packet to be immediately
orphaned. Set /proc/sys/net/bridge/bridge nf-call-iptables to zero to
disable this behaviour.

The sndbuf= parameter is upstream as of this commit:

  http://git.savannah.gnu.org/cgit/qemu.git/commit/?id=0df0ff6de7

I've just now sent a patch upstream to set a default value.

The proposed patch for 5.4 isn't a simple cherry-pick of the code from
upstream, as upstream has an internal API for buffering which would have
been too invasive to backport.

Signed-off-by: Mark McLoughlin <markmc@redhat.com>
Message-Id: <1246356077.3749.34.camel@blaa>
Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>
Bugzilla: 508861
RH-Upstream-status: backported-from-upstream
Acked-by: Herbert Xu <herbert.xu@redhat.com>
Acked-by: Juan Quintela <quintela@redhat.com>
Acked-by: john cooper <john.cooper@redhat.com>
Acked-by: "Michael S. Tsirkin" <mst@redhat.com>